+++ tests/TestLib/Library.fs
@@ -1,0 +1,5 @@
+﻿namespace TestLib
+
+module Test =
+    let Number() =
+        System.Random().Next()